requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
djangorestframework==3.14.0
order==2.1.1
product==0.1.1.1
pytz==2022.7.1
scinum==1.4.4
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
